feat: add api for burger, add hooks
