+++ src/components/burgerConstructor/burgerComponent/burgerComponent.tsx
@@ -1,6 +1,6 @@
 import {DragIcon, ConstructorElement } from "@ya.praktikum/react-developer-burger-ui-components";
 import styles from './burgerComponent.module.css'
-import { BurgerComponentType, IIngredient } from "../../../utils/data";
+import { BurgerComponentType, IIngredient } from "../../api/types";
 
 export interface BurgerComponentProps extends IIngredient {
   positionType?: BurgerComponentType
